Fix infinite loop on parsing spells on arm64. (#42951)

+++ src/magic.cpp
@@ -1050,9 +1050,9 @@ std::string spell::damage_type_string() const
 
 // constants defined below are just for the formula to be used,
 // in order for the inverse formula to be equivalent
-constexpr float a = 6200.0f;
-constexpr float b = 0.146661f;
-constexpr float c = -62.5f;
+constexpr double a = 6200.0;
+constexpr double b = 0.146661;
+constexpr double c = -62.5;
 
 int spell::get_level() const
 {
